changed variable types from int to size_t

+++ misaligned.cpp
@@ -2,8 +2,8 @@
 #include <assert.h>
 #include <string.h>
 
-int getMaxLengthOfString(const char* colorList[]){
-    int maxLengthColor = 0;
+size_t getMaxLengthOfString(const char* colorList[]){
+    size_t maxLengthColor = 0;
     for(size_t i = 0; i < 5; i++)
         if(maxLengthColor < strlen(colorList[i]))
             maxLengthColor = strlen(colorList[i]);
@@ -24,7 +24,7 @@ std::string printWithSpacing(std::string colour, size_t length){
 int printColorMap() {
     const char* majorColor[] = {"White", "Red", "Black", "Yellow", "Violet"};
     const char* minorColor[] = {"Blue", "Orange", "Green", "Brown", "Slate", "Purple"};
-    int maxLengthMajorColor = getMaxLengthOfString(majorColor);
+    size_t maxLengthMajorColor = getMaxLengthOfString(majorColor);
 
     size_t totalColourPair = sizeof(majorColor)/sizeof(majorColor[0]) * sizeof(minorColor)/sizeof(minorColor[0]);
     size_t maxNumberOfDigits = std::to_string(totalColourPair).size();
@@ -43,7 +43,7 @@ int main() {
     const char* majorColourTestList[] = {"Magenta", "Cyan", "Grey", "Orange", "Purple"};
     int result = printColorMap();
     std::string outputString = printWithSpacing("Red" , 5);
-    int lengthOfLongest = getMaxLengthOfString(majorColourTestList);
+    size_t lengthOfLongest = getMaxLengthOfString(majorColourTestList);
     assert(lengthOfLongest == 7);
     assert(outputString.size() == 8);
     assert(outputString[-2] = '|');
